memory leak found on close operation which in turn when fixed highlighted problem with the open / close operations where a statement handle was incorrectly allocated and consequently could crash when freed.  This sequence of events has now been fixed such that connection handles are freed and memory leaks do not occur.

+++ src/OdbcConnectionBridge.cpp
@@ -149,7 +149,7 @@ namespace mssql
 
 	Handle<Value> OdbcConnectionBridge::FreeStatement(Handle<Number> queryId, Handle<Object> callback)
 	{
-		auto id = queryId->IntegerValue();
+		auto id = static_cast<long>(queryId->IntegerValue());
 		nodeTypeFactory fact;
 		auto op = make_shared<FreeStatementOperation>(connection, id, callback);
 		connection->statements->checkin(id);
